Update README.md
The Walking Dead" is a popular American television series that ran from 2010 to 2022. The show follows the story of a group of survivors led by Sheriff Deputy Rick Grimes, who wakes up from a coma to find the world overrun by zombies. The series explores themes of survival, morality, and the challenges of rebuilding society in a post-apocalyptic world.